+++ advent_common/src/intcode/status.rs
@@ -1,6 +1,6 @@
 #[derive(Debug)]
 pub enum Status {
     Exited(anyhow::Result<()>),
-    HasOutput(i32),
+    HasOutput(i64),
     RequiresInput,
 }
